…

+++ src/PetalEventHandler.h
@@ -13,6 +13,6 @@ public:
   void sendSysExMessage(const byte *payload, unsigned payloadLength);
   void sendProgramChange(byte channel, byte number);
   void sendControlChange(byte channel, byte number, byte value);
-  void processPacket(uint32_t data);
+  void processPacket(unsigned long data);
   void onEventFired(float * beat);
 };
